+++ CPUFriend/CPUFriend.cpp
@@ -9,17 +9,21 @@
 
 #include "CPUFriend.hpp"
 
-static const char *kextACPISMC[] = { "/System/Library/Extensions/IOPlatformPluginFamily.kext/Contents/PlugIns/ACPI_SMC_PlatformPlugin.kext/Contents/MacOS/ACPI_SMC_PlatformPlugin" };
-static const char *kextX86PP[] = { "/System/Library/Extensions/IOPlatformPluginFamily.kext/Contents/PlugIns/X86PlatformPlugin.kext/Contents/MacOS/X86PlatformPlugin" };
+static const char *kextACPISMC[] { "/System/Library/Extensions/IOPlatformPluginFamily.kext/Contents/PlugIns/ACPI_SMC_PlatformPlugin.kext/Contents/MacOS/ACPI_SMC_PlatformPlugin" };
+static const char *kextX86PP[] { "/System/Library/Extensions/IOPlatformPluginFamily.kext/Contents/PlugIns/X86PlatformPlugin.kext/Contents/MacOS/X86PlatformPlugin" };
+static const char *kextMCEReporter[] { "/System/Library/Extensions/AppleIntelMCEReporter.kext/Contents/MacOS/AppleIntelMCEReporter" };
+
 
 enum : size_t {
 	KextACPISMC,
-	KextX86PP
+	KextX86PP,
+    KextMCEReporter
 };
 
 static KernelPatcher::KextInfo kextList[] {
 	{ "com.apple.driver.ACPI_SMC_PlatformPlugin", kextACPISMC, arrsize(kextACPISMC), {}, {}, KernelPatcher::KextInfo::Unloaded },
-	{ "com.apple.driver.X86PlatformPlugin", kextX86PP, arrsize(kextX86PP), {}, {}, KernelPatcher::KextInfo::Unloaded }
+	{ "com.apple.driver.X86PlatformPlugin", kextX86PP, arrsize(kextX86PP), {}, {}, KernelPatcher::KextInfo::Unloaded },
+	{ "com.apple.driver.AppleIntelMCEReporter", kextMCEReporter, arrsize(kextMCEReporter), {}, {}, KernelPatcher::KextInfo::Unloaded }
 };
 
 static constexpr size_t kextListSize = arrsize(kextList);
@@ -38,7 +42,7 @@ IOService *CPUFriendData::probe(IOService *provider, SInt32 *score)
 				if (!data) {
 					auto cpu = provider->getParentEntry(gIOServicePlane);
 					if (cpu) {
-						DBGLOG("cpuf", "looking for cf-frequency-data in %s", safeString(cpu->getName()));
+						DBGLOG("cpuf", "looking for cf-frequency-data again in %s", safeString(cpu->getName()));
 						data = OSDynamicCast(OSData, cpu->getProperty("cf-frequency-data"));
 					} else {
 						SYSLOG("cpuf", "unable to access cpu parent");
@@ -61,16 +65,14 @@ IOService *CPUFriendData::probe(IOService *provider, SInt32 *score)
 	return nullptr;
 }
 
-bool CPUFriendPlugin::init()
+void CPUFriendPlugin::init()
 {
 	callbackCpuf = this;
 	
 	lilu.onKextLoadForce(kextList, kextListSize,
-	[](void *user, KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size) {
-		static_cast<CPUFriendPlugin *>(user)->processKext(patcher, index, address, size);
-	}, this);
-
-	return true;
+						 [](void *user, KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size) {
+							static_cast<CPUFriendPlugin *>(user)->processKext(patcher, index, address, size);
+						 }, this);
 }
 
 void CPUFriendPlugin::updateResource(kern_return_t &result, const void * &resourceData, uint32_t &resourceDataLength)
@@ -84,7 +86,11 @@ void CPUFriendPlugin::updateResource(kern_return_t &result, const void * &resour
 			resourceDataLength = sz;
 			result = kOSReturnSuccess;
 		} else {
-			SYSLOG("cpuf", "failed to feed cpu data (%u, %d)", sz, data != nullptr);
+			// this is fine when not providing customized data,
+			// in the worst case it's just the original
+			// frequencyData and frequencyDataSize get handled,
+			// which looks safe enough.
+			DBGLOG("cpuf", "failed to feed cpu data (%u, %d), feeding data from org callback", sz, data != nullptr);
 		}
 	} else {
 		SYSLOG("cpuf", "config callback arrived at nowhere");
@@ -107,6 +113,14 @@ void CPUFriendPlugin::myX86PPConfigResourceCallback(uint32_t requestTag, kern_re
 	FunctionCast(myX86PPConfigResourceCallback, callbackCpuf->orgX86PPConfigLoadCallback)(requestTag, result, resourceData, resourceDataLength, context);
 }
 
+IOService *CPUFriendPlugin::myAppleIntelMCEReporterProbe(IOService *, IOService *, SInt32 *)
+{
+	DBGLOG("cpuf", "killing AppleIntelMCEReporter");
+
+	// always return nullptr for prevention of AppleIntelMCEReporter.
+	return nullptr;
+}
+
 void CPUFriendPlugin::processKext(KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size)
 {
 	if (kextList[KextACPISMC].loadIndex == index) {
@@ -122,4 +136,12 @@ void CPUFriendPlugin::processKext(KernelPatcher &patcher, size_t index, mach_vm_
 											orgX86PPConfigLoadCallback);
 		patcher.routeMultiple(index, &request, 1, address, size);
 	}
+    
+	if (kextList[KextMCEReporter].loadIndex == index) {
+		DBGLOG("cpuf", "patching AppleIntelMCEReporter");
+		KernelPatcher::RouteRequest request("__ZN21AppleIntelMCEReporter5probeEP9IOServicePi",
+											myAppleIntelMCEReporterProbe);
+		patcher.routeMultiple(index, &request, 1, address, size);
+	}
+
 }