+++ Changelog.md
@@ -2,6 +2,7 @@ CPUFriend Changelog
 ===================
 #### v1.1.9
 - Unified release archive names
+- Added support for prevention of AppleIntelMCEReporter, details can be found [here](https://github.com/acidanthera/bugtracker/issues/424#issuecomment-512596034)
 
 #### v1.1.8
 - Allow loading on 10.15 without `-lilubetaall`